First version of job runner that can run local tests of quants end to end. Lots of things renamed/made public so the new jobmanager can access it, eg parts of quantification package. The new tests don't rely on s3 buckets but can test in local file system using FS mocking. Other things moved to break circular references. Executor removed as it seems no longer needed with new JobNode code (that handles starting docker containers on an EC2 for eg)

+++ api/quantification/quantRunner/nodeBalance.go
@@ -22,18 +22,18 @@ func EstimateNodeCount(spectraCount uint, elementCount uint, desiredRunTimeSec u
 	// See unit test for why...
 
 	// Add 0.5 to round up, can't have it fractional
-	nodeCount := uint((float32(spectraCount*(elementCount+3)) / float32(3*desiredRunTimeSec*coresPerNode)) + 0.5)
+	nodeCount := int((float32(spectraCount*(elementCount+3)) / float32(3*desiredRunTimeSec*coresPerNode)) + 0.5)
 
 	// Clamp it to reasonable values
 	if nodeCount < 1 {
 		nodeCount = 1
 	}
 	// Don't go way overboard either
-	if maxNodes > 0 && nodeCount > maxNodes {
-		nodeCount = maxNodes
+	if maxNodes > 0 && nodeCount > int(maxNodes) {
+		nodeCount = int(maxNodes)
 	}
 
-	return nodeCount
+	return uint(nodeCount)
 }
 
 func FilesPerNode(spectraCount uint, nodeCount uint) uint {
